Add files via upload
The Tic-Tac-Toe Project is a classic and timeless game brought to life in the digital realm, offering players a fun and strategic gaming experience, implemented using web technologies like HTML, CSS, and Javascript. This project brings together simplicity and entertainment, making it accessible to players of all ages. Whether you're a casual gamer looking for a quick match or a seasoned strategist seeking a challenge, Tic-Tac-Toe has something to offer. It allows two players to go head-to-head in a battle of Xs and Os.